IS-3400: Legg graf inn i egen komponent
